Add provider-specific options, make providers re-orderable
As suggested by @Mineo in https://github.com/musicbrainz/picard/pull/411#issuecomment-118076874

- add an option subpage per provider having options
- in the main cover art options page display a list of checkboxes with up/down arrows for re-ordering
- code is simplified
- more space is available for provider-specific options
- no drag'n'drop for re-ordering
