+++ picard/ui/ui_options_cover.py
@@ -22,7 +22,7 @@ except AttributeError:
 class Ui_CoverOptionsPage(object):
     def setupUi(self, CoverOptionsPage):
         CoverOptionsPage.setObjectName(_fromUtf8("CoverOptionsPage"))
-        CoverOptionsPage.resize(595, 560)
+        CoverOptionsPage.resize(632, 560)
         self.verticalLayout = QtGui.QVBoxLayout(CoverOptionsPage)
         self.verticalLayout.setObjectName(_fromUtf8("verticalLayout"))
         self.rename_files = QtGui.QGroupBox(CoverOptionsPage)
@@ -50,77 +50,23 @@ class Ui_CoverOptionsPage(object):
         self.save_images_overwrite.setObjectName(_fromUtf8("save_images_overwrite"))
         self.vboxlayout.addWidget(self.save_images_overwrite)
         self.verticalLayout.addWidget(self.rename_files)
-        self.groupBox = QtGui.QGroupBox(CoverOptionsPage)
-        self.groupBox.setObjectName(_fromUtf8("groupBox"))
-        self.verticalLayout_2 = QtGui.QVBoxLayout(self.groupBox)
-        self.verticalLayout_2.setObjectName(_fromUtf8("verticalLayout_2"))
-        self.caprovider_amazon = QtGui.QCheckBox(self.groupBox)
-        self.caprovider_amazon.setObjectName(_fromUtf8("caprovider_amazon"))
-        self.verticalLayout_2.addWidget(self.caprovider_amazon)
-        self.caprovider_caa = QtGui.QCheckBox(self.groupBox)
-        self.caprovider_caa.setObjectName(_fromUtf8("caprovider_caa"))
-        self.verticalLayout_2.addWidget(self.caprovider_caa)
-        self.caprovider_whitelist = QtGui.QCheckBox(self.groupBox)
-        self.caprovider_whitelist.setObjectName(_fromUtf8("caprovider_whitelist"))
-        self.verticalLayout_2.addWidget(self.caprovider_whitelist)
-        self.verticalLayout.addWidget(self.groupBox)
-        self.gb_caa = QtGui.QGroupBox(CoverOptionsPage)
-        self.gb_caa.setEnabled(False)
-        self.gb_caa.setObjectName(_fromUtf8("gb_caa"))
-        self.verticalLayout_3 = QtGui.QVBoxLayout(self.gb_caa)
-        self.verticalLayout_3.setObjectName(_fromUtf8("verticalLayout_3"))
-        self.horizontalLayout = QtGui.QHBoxLayout()
-        self.horizontalLayout.setObjectName(_fromUtf8("horizontalLayout"))
-        self.label = QtGui.QLabel(self.gb_caa)
-        self.label.setObjectName(_fromUtf8("label"))
-        self.horizontalLayout.addWidget(self.label)
+        self.ca_providers_groupbox = QtGui.QGroupBox(CoverOptionsPage)
+        sizePolicy = QtGui.QSizePolicy(QtGui.QSizePolicy.Preferred, QtGui.QSizePolicy.Preferred)
+        sizePolicy.setHorizontalStretch(0)
+        sizePolicy.setVerticalStretch(0)
+        sizePolicy.setHeightForWidth(self.ca_providers_groupbox.sizePolicy().hasHeightForWidth())
+        self.ca_providers_groupbox.setSizePolicy(sizePolicy)
+        self.ca_providers_groupbox.setObjectName(_fromUtf8("ca_providers_groupbox"))
+        self.ca_providers_layout = QtGui.QVBoxLayout(self.ca_providers_groupbox)
+        self.ca_providers_layout.setObjectName(_fromUtf8("ca_providers_layout"))
+        self.ca_providers_list = QtGui.QHBoxLayout()
+        self.ca_providers_list.setObjectName(_fromUtf8("ca_providers_list"))
         spacerItem = QtGui.QSpacerItem(40, 20, QtGui.QSizePolicy.Expanding, QtGui.QSizePolicy.Minimum)
-        self.horizontalLayout.addItem(spacerItem)
-        self.cb_image_size = QtGui.QComboBox(self.gb_caa)
-        sizePolicy = QtGui.QSizePolicy(QtGui.QSizePolicy.Preferred, QtGui.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.cb_image_size.sizePolicy().hasHeightForWidth())
-        self.cb_image_size.setSizePolicy(sizePolicy)
-        self.cb_image_size.setObjectName(_fromUtf8("cb_image_size"))
-        self.cb_image_size.addItem(_fromUtf8(""))
-        self.cb_image_size.addItem(_fromUtf8(""))
-        self.cb_image_size.addItem(_fromUtf8(""))
-        self.horizontalLayout.addWidget(self.cb_image_size)
-        self.verticalLayout_3.addLayout(self.horizontalLayout)
-        self.select_caa_types_group = QtGui.QHBoxLayout()
-        self.select_caa_types_group.setObjectName(_fromUtf8("select_caa_types_group"))
-        self.restrict_images_types = QtGui.QCheckBox(self.gb_caa)
-        self.restrict_images_types.setObjectName(_fromUtf8("restrict_images_types"))
-        self.select_caa_types_group.addWidget(self.restrict_images_types)
-        spacerItem1 = QtGui.QSpacerItem(40, 20, QtGui.QSizePolicy.Expanding, QtGui.QSizePolicy.Minimum)
-        self.select_caa_types_group.addItem(spacerItem1)
-        self.select_caa_types = QtGui.QPushButton(self.gb_caa)
-        sizePolicy = QtGui.QSizePolicy(QtGui.QSizePolicy.Fixed, QtGui.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(100)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.select_caa_types.sizePolicy().hasHeightForWidth())
-        self.select_caa_types.setSizePolicy(sizePolicy)
-        self.select_caa_types.setObjectName(_fromUtf8("select_caa_types"))
-        self.select_caa_types_group.addWidget(self.select_caa_types)
-        self.verticalLayout_3.addLayout(self.select_caa_types_group)
-        self.cb_approved_only = QtGui.QCheckBox(self.gb_caa)
-        self.cb_approved_only.setObjectName(_fromUtf8("cb_approved_only"))
-        self.verticalLayout_3.addWidget(self.cb_approved_only)
-        self.cb_type_as_filename = QtGui.QCheckBox(self.gb_caa)
-        sizePolicy = QtGui.QSizePolicy(QtGui.QSizePolicy.Minimum, QtGui.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.cb_type_as_filename.sizePolicy().hasHeightForWidth())
-        self.cb_type_as_filename.setSizePolicy(sizePolicy)
-        self.cb_type_as_filename.setObjectName(_fromUtf8("cb_type_as_filename"))
-        self.verticalLayout_3.addWidget(self.cb_type_as_filename)
-        self.caprovider_caa_release_group = QtGui.QCheckBox(self.gb_caa)
-        self.caprovider_caa_release_group.setObjectName(_fromUtf8("caprovider_caa_release_group"))
-        self.verticalLayout_3.addWidget(self.caprovider_caa_release_group)
-        self.verticalLayout.addWidget(self.gb_caa)
-        spacerItem2 = QtGui.QSpacerItem(275, 91, QtGui.QSizePolicy.Minimum, QtGui.QSizePolicy.Expanding)
-        self.verticalLayout.addItem(spacerItem2)
+        self.ca_providers_list.addItem(spacerItem)
+        self.ca_providers_layout.addLayout(self.ca_providers_list)
+        self.verticalLayout.addWidget(self.ca_providers_groupbox)
+        spacerItem1 = QtGui.QSpacerItem(20, 40, QtGui.QSizePolicy.Minimum, QtGui.QSizePolicy.Expanding)
+        self.verticalLayout.addItem(spacerItem1)
 
         self.retranslateUi(CoverOptionsPage)
         QtCore.QObject.connect(self.save_images_to_tags, QtCore.SIGNAL(_fromUtf8("clicked(bool)")), self.cb_embed_front_only.setEnabled)
@@ -129,14 +75,6 @@ class Ui_CoverOptionsPage(object):
         CoverOptionsPage.setTabOrder(self.cb_embed_front_only, self.save_images_to_files)
         CoverOptionsPage.setTabOrder(self.save_images_to_files, self.cover_image_filename)
         CoverOptionsPage.setTabOrder(self.cover_image_filename, self.save_images_overwrite)
-        CoverOptionsPage.setTabOrder(self.save_images_overwrite, self.caprovider_amazon)
-        CoverOptionsPage.setTabOrder(self.caprovider_amazon, self.caprovider_caa)
-        CoverOptionsPage.setTabOrder(self.caprovider_caa, self.caprovider_whitelist)
-        CoverOptionsPage.setTabOrder(self.caprovider_whitelist, self.cb_image_size)
-        CoverOptionsPage.setTabOrder(self.cb_image_size, self.select_caa_types)
-        CoverOptionsPage.setTabOrder(self.select_caa_types, self.cb_approved_only)
-        CoverOptionsPage.setTabOrder(self.cb_approved_only, self.cb_type_as_filename)
-        CoverOptionsPage.setTabOrder(self.cb_type_as_filename, self.caprovider_caa_release_group)
 
     def retranslateUi(self, CoverOptionsPage):
         self.rename_files.setTitle(_("Location"))
@@ -145,18 +83,5 @@ class Ui_CoverOptionsPage(object):
         self.save_images_to_files.setText(_("Save cover images as separate files"))
         self.label_3.setText(_("Use the following file name for images:"))
         self.save_images_overwrite.setText(_("Overwrite the file if it already exists"))
-        self.groupBox.setTitle(_("Coverart Providers"))
-        self.caprovider_amazon.setText(_("Amazon"))
-        self.caprovider_caa.setText(_("Cover Art Archive"))
-        self.caprovider_whitelist.setText(_("Sites on the whitelist"))
-        self.gb_caa.setTitle(_("Cover Art Archive"))
-        self.label.setText(_("Only use images of the following size:"))
-        self.cb_image_size.setItemText(0, _("250 px"))
-        self.cb_image_size.setItemText(1, _("500 px"))
-        self.cb_image_size.setItemText(2, _("Full size"))
-        self.restrict_images_types.setText(_("Download only cover art images matching selected types"))
-        self.select_caa_types.setText(_("Select types..."))
-        self.cb_approved_only.setText(_("Download only approved images"))
-        self.cb_type_as_filename.setText(_("Use the first image type as the filename. This will not change the filename of front images."))
-        self.caprovider_caa_release_group.setText(_("Use the image of the release group if no front image is associated with the release"))
+        self.ca_providers_groupbox.setTitle(_("Cover Art Providers"))
